New script-based architecture for filters.
All filters now have corresponding scripting interfaces.
The dialogs in the GUI that previously applied filters
directly have been changed to generate scripting commands
instead: each generated command is executed in order to
apply the filter, and then added to the history log.  The
history is now saved as the script "~/.rheat/history.js".

Since scripting commands tend to be long and helix info
tends to be narrow, the window layout has been changed.
The script-based history log is at the bottom (as wide as
the display window) and helix info is now on the left.