Improved rendering and behavior of zooming.
Image scaling has been replaced by a transformation on
the graphics context itself, allowing zoomed-in helix
lines to be rendered at high quality instead of having
a pixelated appearance.  The code for generating an
image has been separated from the code that draws the
content in preparation for future changes to speed up
drawing.  (An image is rarely required; it is better
to draw directly into a context from a component, and
avoid massive memory allocations.)

Also, the display window now remains centered around
the same area after being zoomed in or out.