Usability improvements to GUI.
Removed over 300 lines of code by changing the way dialogs
are displayed (such as Preferences and filters), relying on
JOptionPane instead of having custom buttons and panels.
This also means that keys like Return and Escape will work.

Now fixed-width font is used for the info and history panes.
This makes base-pair strings line up properly, and it makes
the history commands easier to read.