Filled in help content and improved navigation.
All of the HTML help content has now been written.  In addition,
the help window has been updated to recognize clicks on links and
to support the ability to go back to previously-viewed pages.
Menus have been tweaked slightly, to include the new "Go Back"
mechanism under Help and to include an explicit Undo command
(with a standard key equivalent that uses 'Z').