Fixed some window layering and behavior issues.
The “Helix Info” and Help windows are now placed on the
palette layer, which keeps them above all the other windows.
This behavior is more convenient when the display has been
zoomed to full screen.

Fixed an exception that could occur when clicking on the
helix display while no file is loaded.