Added script support for helix iteration.
It is now possible to use a script to iterate over actual
helices (upper-right of diagonal) or predicted helices
(lower-left of diagonal, affected by filters).  Also added
a way to find any currently-selected helix.  Each helix
can be queried for properties such as length and energy.
Included a test case to show behavior, and updated the
help content accordingly.