Zoom controls have been improved.
Zoom buttons have moved to the top of the display,
and the menu has been replaced by a slider for
continuous zoom.  There is also a new option to
zoom to “fit”, which automatically calculates the
magnification to show the entire RNA in the window.
The pop-up busy-window has been removed, replaced
by a mouse wait-cursor (as this was distracting
when using the new continuous-zoom slider).