Starting to rename “filter” to “constraint”.
As suggested, references to the term “filter” are starting to
be replaced by “constraint”.  This commit focuses on elements
of the GUI but it does not change the names of scripting
commands or Java source files.  (A future commit should also
update those elements.)  Also, constraints still act like
filters; a future change will allow their effects to generate
overlays instead of removing helices from the display.