Fixed inconsistencies in actual-helix data.
Modified the helix data definition so that all helices have
the same base origin regardless of where they appear in the
grid, and fixed an off-by-one error in the parser.  These
fixes ensure that numbers are correct for selected helices,
and helices on either side of the diagonal are easier to
compare.  The offset of “actual” helices is now decided by
the image generator, and not the data in the helix store.
Also fixed minor problem where brackets on selected helices
could be shifted.