Added grid lines/preferences; image bug fixes.
There is now a faint grid displayed in the background
of the RNA display.  The spacing of the lines can be
customized with a new preference setting.  (Also took
the opportunity to update the layout of the Preferences
dialog, and the appearance of other dialog frames.)
Maximizing/restoring the display window will now
automatically zoom-to-fit the content.  Fixed saved
images so they have a white background.