Added mini-frame to RNA display.
The mini-frame appears only when some part of the
RNA is not visible in the scroll view (typically
because it has been zoomed in to some degree).
Currently the mini-frame is just a dark gray square
with a white rectangle representing the region that
is visible.  The mini-frame is in a movable window
that is ordered above all other windows.