Added diagonal scrolling buttons.
In the lower-right corner when scroll bars are visible,
there are now two buttons for simultaneously scrolling
in both directions (upper-left and lower-right).  For
now, these occupy the available space in the corner so
the buttons are quite small.  In the future, may be
able to find a way to make them larger.