Added annotations to the graphical interface.
A new command in the File menu allows annotations to be
read from ".bpcolor" files (same as the recent addition
for doing this from a script).  When the annotations
are read, the display is refreshed and annotated helices
appear in a different color with a slightly thicker line.
Preferences have been extended to set a single default
color for annotated helices, different from the normal
helix color.  There are plans to extend preferences to
allow particular tags to have their own unique colors,
and to provide a way to show/hide helix annotations by
name; for now, only basic annotation support exists.