Added grid labels, fast scroll and window layers.
There are now periodic labels along the diagonal, showing the
corresponding X/Y coordinates of each position on the grid.
Scrolling rates have been adjusted so that it is easier to
move around the display by clicking on arrows or buttons.
All non-main windows are now layered on top so that they will
not fall behind the main display.  It is still possible to
minimize windows to keep them out of the way.