Added customization of line thickness by tag.
Similar to the script command setHelixTagColor(), there
is now setHelixTagLineWidth() to customize the thickness
of the line used for an annotated helix.  As with colors
the line thickness depends on the priority and visibility
of the tags that are on a particular helix.

Also, the spectrum display is now off by default so that
the initial display is not excessively annotated with
dummy energy values.