Recently-run scripts now available in File menu.
When a script is run using the “Run Script…” command, that
script is now added to the File menu for reuse in the
current session (in the order it was first used).  Also,
there is now a wait cursor displayed just before scripts
and programs are run from the menu, in case they take some
time to finish.