Added color editor and file browser to console.
For convenience, the scripting console now has a button
next to the current working directory that can be used
to browse for a new path and automatically generate and
run the equivalent command to switch to that directory.
The new command also appears in the scripting history.

The console also now has a color editor.  The displayed
color is not used but the editor helps when generating
color strings for scripting commands.  Copy and Paste
are allowed in the color field.