Clarifications to helix-store interface.
Changed names of some methods in HelixStore/HelixGrid to
make it more obvious what they are referring to.  Also,
extended the interface with some minor methods like
isEmpty().  Soon, should add an interface for returning
a bounded iterator, since it is usually important to be
able to focus on a subset of helices.