Major update to use tags instead of filters.
Constraints no longer have filter-like behavior; they apply
special tag names to matching helices.  (In order to hide
helices that do not match any constraint or any other form
of annotation, the recently-added checkbox can be clicked to
hide untagged helices.)

This opens up a lot of flexibility; for one, all scripting
commands that allow customization of tags to show/hide,
colorize, set line width, etc. will work for the special tag
names that are now used by constraints.  (The tag names can
be seen by clicking on helices that match a constraint.)

This approach can also be quite a bit faster: constraints no
longer create copies of RNAs, they tag the originals and the
Undo operation is implemented just by reversing the tagging.

One implication of the lack of real filters is that iteration
now returns all helices, all the time.  To exclude helices
that would have matched a constraint, it is now necessary to
check for matching tags.  The method hasTag() can be used for
this, as shown in the test-case examples.  Of course, this
can also be used to do much more sophisticated filtering.
If performance is an issue, more iteration options should be
created to select subsets of helices in advance.