Added key-value tags, and improved performance.
The ".helixcolor" format now allows "key=value" forms
for tags.  The key portion is treated the same as any
other tag; a new scripting API has been added to find
tag values.  A tag-only case is a null value.

The performance of processing annotations has greatly
improved because it is now possible to iterate over a
subset of helices.  In the case of tags, checks on
perhaps hundreds of thousands of helices can be
reduced down to a few.

Built-in tags from constraints now take advantage of
the tag-value portion to communicate parameters (such
as the maximum and minimum boundaries of a length
constraint).  This could be read from scripts and is
also useful to remind the user why a tag is present.