Improved usability of text fields.
Editable text areas now use a modified focus scheme that
provides more padding and very obvious border styles.
Also, single-line fields now select all of their text if
the user has tabbed to the field using the keyboard (not
the mouse, as the user may wish to highlight a subset of
the text).  This makes it very easy to replace old values
with new ones.