Improved data model and selected-helix display.
The helix model has been significantly simplified,
removing direct storage of X or Y values and relying
entirely on 5' and 3' ranges.  This makes it much
easier to tell what is happening, and it eliminates
some error-prone translation steps.  While making
these changes, a bug in the scripting API was fixed
so that 5' and 3' start/end values are no longer off
by one (compared to numbers in the GUI display).
The graphical “selected helix” display has been
updated as well: diagonal distance is now included,
tags are now sorted, and the layout has been changed
to make properties easier to read.