Added a removeTag() Helix scripting API.
Similar to addTag(), removeTag() operates on an
individual helix (such as in an iteration loop)
to delete annotations.  As in addTag(), the
effects are NOT immediately seen on the display;
the user must trigger something explicit on the
entire RNA (such as, "rheat.showHelixTag(...)")
to cause the display to be updated.

Also, fixed a problem with an exception if a
delay was invoked on the event dispatch thread.